Bubble short
The Bubble Sort algorithm is a simple sorting algorithm that repeatedly steps through the list, compares adjacent elements, and swaps them if they are in the wrong order. Here's how it works:

>Iterate Through the List: Start from the beginning of the list/array.
>Compare Adjacent Elements: Compare each pair of adjacent elements in the list.
>Swap if Necessary: If the elements are in the wrong order (i.e., the first element is greater than the second), swap them.
>Repeat Until Sorted: Continue this process for each pair of adjacent elements in the list, moving from the beginning to the end. After the first iteration, the largest element 
    will be in its correct position at the end of the list. Repeat this process for the remaining elements, excluding the last one which is already sorted.
>Continue Iterating: Continue iterating and swapping until no more swaps are needed, indicating that the list is sorted.


Bubble Sort is not very efficient for large lists, as its average and worst-case time complexity is O(n^2), where n is the number of elements in the list. However, it is simple to understand and implement, making it useful for small lists or educational purposes.